pyproject.toml
[tool.poetry]
name = "issuer-service"
version = "0.1.0"
description = "Issuing NFTs on Liquid"
authors = ["Jürgen Eckel <juergen@riddleandcode.com>"]
license = "AGPLv3"
readme = "README.md"
packages = [{include = "issuer_service"}]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
python-bitcoinrpc = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
